refactor: move request button folder
